Step 1 of dynamic command refactor.
