+++ commands/command.go
@@ -1,0 +1,8 @@
+package commands
+import (
+	"github.com/mattermost/mattermost-server/v5/model"
+)
+
+type Command struct {
+    event *model.WebSocketEvent
+}
